Add new homepage layout and components
